requirements.txt
# Install the shared core package separately:
#   pip install -e ../paired-single-cell-pipeline

adjusttext>=1.3.0
anndata>=0.11.0
celltypist>=1.7.0
harmonypy>=0.2.0
h5py>=3.16.0
igraph>=1.0.0
leidenalg>=0.11.0
matplotlib>=3.10.0
numpy>=2.2.0
pandas>=2.3.0
pyyaml>=6.0.0
scanpy>=1.11.0
scipy>=1.15.0
seaborn>=0.13.0
statsmodels>=0.14.0
tables>=3.10.0
typer>=0.12.0
umap-learn>=0.5.0

